(1)Created migration files for project app (2)modified project page (3)fixed minor bugs

+++ project/views.py
@@ -39,7 +39,7 @@ def show_project_page(request,id):
 		item.backer=len(pledger)
 		for key,value in pledger.iteritems():
 			if not value is None:
-				item.percent=(100*(float(value)/float(item.money_req)))
+				item.percent=int(100*(float(value)/float(item.money_req)))
 				item.amount_pledged=value
 			else:
 				item.percent=0
